massive structure change... again...
