+++ marconi/__init__.py
@@ -1,10 +1,4 @@
 # core
 import logging
-# 3rd party
-from twisted.internet import reactor
-from twisted.internet.defer import inlineCallbacks
-from autobahn.twisted.wamp import ApplicationSession, ApplicationRunner
-# local
-from tools.minion import Minion
-# logger
+
 logger = logging.getLogger(__name__)
