+++ marconi/__main__.py
@@ -1,9 +1,5 @@
-# pip install daemon
-from daemon import Daemon
-# https://github.com/s4w3d0ff/python-poloniex
-from poloniex import Poloniex, Coach
-# local
-from bots import loaner, liquidator, lurker, ticker
+#!/usr/bin/python
+from tools import Daemon, Poloniex  # , Coach
 
 PIDPATH = '~/.marconid.pid'
 
@@ -16,20 +12,19 @@ class Marconid(Daemon):
         super(Daemon, self).__init__(*args, **kwargs)
         self.api = Poloniex(coach=True, jsonNums=float)
         self.bots = {}
-        self.bots['loaner'] = loaner.Loaner(self.api)
-        self.bots['lurker'] = lurker.Lurker(self.api)
-        self.bots['ticker'] = ticker.Ticker(self.api)
-        self.bots['911'] = liquidator.Liquidator(self.api)
 
     def run(self):
         while True:
             for bot in bots:
-                if not bots[bot]._running:
-                    bots[bot].start()
+                logging.info("%s Running: %s", bot, str(bots[bot]._running))
+
+    def watch(self, market):
+
 
 if __name__ == '__main__':
+    import sys
     d = Watcher(PIDPATH)
     command = str(sys.argv[1]).lower()
     if command not in COMMANDS:
-        raise Exception('Invalid command: %s' % arg)
-    getattr(spawn, command)(*argv[2:])
+        raise Exception('Invalid command: %s' % command)
+    getattr(d, command)(sys.argv[2:])
